학생 오답노트

+++ realProject2/src/main/java/icia/project/dao/IMybatis.java
@@ -187,16 +187,16 @@ public interface IMybatis {
 	public int learningDataDelete(BoardBean board);
 
 	// 학습방 오답문제 총 평균 구하기1 (학년 총 질문수)
-	public int learningWANstAverage1(BoardBean board);
+	public double learningWANstAverage1(BoardBean board);
 
 	// 학습방 오답문제 총 평균 구하기2 (학년 총 인원수) 
-	public int learningWANstAverage2(BoardBean board);
+	public double learningWANstAverage2(BoardBean board);
 
 	// 학습방 오답문제 총 평균 구하기3 ( 방 학년 총 질문수)
-	public int learningWANstAverage3(BoardBean board);
+	public double learningWANstAverage3(BoardBean board);
 
 	// 학습방 오답문제 총 평균 구하기4 ( 방 학년 총 인원수)
-	public int learningWANstAverage4(BoardBean board);
+	public double learningWANstAverage4(BoardBean board);
 
 	// 학습방 오답문제 총 평균 구하기5 학생(본인) 오답노트 추출 (중복제외)
 	public ArrayList<BoardBean> learningWANstListGetOverlap(BoardBean board);
